requirements.txt
biopython>=1.79
pandas>=1.3.0
numpy>=1.20.0
tqdm>=4.60.0
